third_party/blink/web_tests/platform/generic/fast/mediastream/MediaStream-stop-expected.txt

Test cloning mediastreams onactive/oninactive callbacks.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS cloned_stream.active is true
PASS cloned_stream.getVideoTracks().length is 1
PASS cloned_stream.getAudioTracks().length is 1
PASS successfullyParsed is true

TEST COMPLETE
